Code cleanup and addition of cors headers
